Refactoring to hopefully make this easier to follow, fewer single-use functions
